+++ packages/runtime-js/src/module-loader/esm-job-types.ts
@@ -44,7 +44,7 @@ export type EsmFactory = (
   metaDirname: string,
   metaFilename: string,
   assetPath: (s: string) => string,
-  metaResolve: (s: string) => string,
+  metaResolve: (s: string, ...parent: unknown[]) => string,
   Function: FunctionConstructor,
   webAssembly: typeof WebAssembly,
   globalKeyCheck: (key: unknown) => unknown,